+++ CHANGELOG.md
@@ -8,8 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.x.y] - Next version
 
 * List users
+* List and delete realms
 * Support passing rest client options for user save and search
-* List realms
 * Support using gem without ActiveSupport
 
 ## [0.7.0] - 2019-06-06